requirements.txt
azure-storage-blob
python-dotenv
requests